Array task 2 done.

+++ Array-tasks/Array-task2/Array-task2/Program.cs
@@ -9,7 +9,7 @@ namespace Array_task2
         {
             Console.WriteLine("Ohjelma arpoo sata lukua väliltä 0-50. Arvotut luvut näytetään ja keskiarvo kerrotaan käyttäjälle.\n");
 
-            int[] numbers = new int[100];
+            double[] numbers = new double[100];
             Random rnd = new Random();
 
             for (int i = 0; i < numbers.Length; i++)
@@ -17,8 +17,8 @@ namespace Array_task2
                 numbers[i] = rnd.Next(50);
             }
 
-            int sum = numbers.Sum();
-            int average = sum / 100;
+            double sum = numbers.Sum();
+            double average = sum / 100;
 
             for (int i = 0; i < numbers.Length; i++)
             {
